requirements.txt
# Core project/common requirements
celery==3.1.18
Django==1.8.2
django-appconf==1.0.1
django-bootstrap3==5.4.0
django-compressor==1.5
django-haystack==2.3.1
djangorestframework==3.1.3
dj-database-url==0.3.0
dj-static==0.0.6
elasticsearch==1.5.0
git+https://github.com/bpeschier/django-compressor-requirejs@889d5edc4f2acaa961c170084f1171c700649519#egg=githubdjango-compressor-requirejs-develop
psycopg2==2.6
PyYAML==3.11
redis==2.10.3
static3==0.5.1
tox>=2.0.2,<3.0.0
uwsgi==2.0.10

# django guardian branch that fixed bug in master
git+https://github.com/mitodl/django-guardian@de7ca5c39dfd0870067497f5843581e9cd2b017f#egg=django-guardian-dev


# cas application requirements
git+https://github.com/mingchen/django-cas-ng@69463fe7c23f025be7165c2967f4ceb983d60472#egg=django-cas-ng

# Importer application requirements
git+https://github.com/mitodl/xbundle.git@7799a0957c502103fb030fc0b16c64204f551c75#egg=xbundle
git+https://github.com/gdub/python-archive.git@61b9afde21621a8acce964d3336a7fb5d2d07ca6#egg=python-archive
django-storages-redux==1.2
python-magic==0.4.6
boto>=2.38.0,<3.0.0
